services/analytics-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
redis==5.0.1
celery==5.3.4
pika==1.3.2

pytest==7.4.3
pytest-asyncio==0.21.1
fakeredis==2.20.0
